ref: brakeEfficiency float; rem: pitch from static tractive effort

+++ src/main/java/cam72cam/immersiverailroading/entity/EntityMoveableRollingStock.java
@@ -452,7 +452,7 @@ public abstract class EntityMoveableRollingStock extends EntityCustomPlayerMovem
      * That percentage was improved over time, with the materials used (friction coefficient) helping inform our guess
      * Though, I'm going to limit it to 75% of the total possible adhesion
      */
-    public double getBrakeSystemEfficiency() {
+    public float getBrakeSystemEfficiency() {
         return getDefinition().getBrakeShoeFriction();
     }
