+++ src/main/java/cam72cam/immersiverailroading/registry/EntityRollingStockDefinition.java
@@ -1043,9 +1043,10 @@ public abstract class EntityRollingStockDefinition {
         return tiltMultiplier;
     }
 
-    public double getBrakeShoeFriction() {
+    public float getBrakeShoeFriction() {
         return brakeCoefficient;
     }
+    
     public int getSnowLayers() {
         return snowLayers;
     }
